Updates and renaming.
* Change asset names for generic workshop name
* Update instruction website location
* Upgrade AWS Encryption SDK for Python
