+++ NOTICE.txt
@@ -1,4 +1,4 @@
-reinvent-sid345-workshop-sample
+busy-engineers-encryption-sdk
 Copyright 2017 Amazon.com, Inc. or its affiliates. All rights reserved.
 
 **********************
