+++ deploy.config
@@ -1,7 +1,7 @@
 [deploy]
-region: eu-west-1
-stack_name: reinvent-sid345-python
-lambda_s3_key: reinvent-sid345-python.zip
+region: us-east-2
+stack_name: busy-engineers-workshop-python
+lambda_s3_key: busy-engineers-workshop-python.zip
 
 [resources]
 template: cloudformation.yaml
