+++ setup.cfg
@@ -17,7 +17,7 @@ ignore_missing_imports = True
 max_complexity = 10
 max_line_length = 120
 import_order_style = google
-application_import_names = reinvent_sid345
+application_import_names = busy_engineers_workshop
 builtins = raw_input
 ignore =
     # E203 is not PEP8 compliant https://github.com/ambv/black#slices
@@ -36,5 +36,5 @@ include_trailing_comma = True
 force_grid_wrap = 0
 combine_as_imports = True
 not_skip = __init__.py
-known_first_party = reinvent_sid345
-known_third_party =aws_encryption_sdk,boto3,reinvent_sid345,setuptools
+known_first_party = busy_engineers_workshop
+known_third_party =aws_encryption_sdk,boto3,busy_engineers_workshop,setuptools
