+++ src/busy_engineers_workshop/__init__.py
@@ -10,4 +10,4 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-"""Python 3.6 Lambda for SID345 Re:Invent workshop."""
+"""Python 3.6 Lambda for Busy Engineer's Guide to the AWS Encryption SDK workshop."""
