+++ tox.ini
@@ -30,7 +30,7 @@ envlist =
 # build :: Builds source and wheel dist files.
 
 [testenv:base-command]
-commands = pytest --basetemp={envtmpdir} -l --cov reinvent_sid345 test/ {posargs}
+commands = pytest --basetemp={envtmpdir} -l --cov busy_engineers_workshop test/ {posargs}
 
 [testenv]
 passenv =
@@ -68,7 +68,7 @@ deps = {[testenv:mypy-common]deps}
 commands =
     python -m mypy \
         --linecoverage-report build \
-        src/reinvent_sid345/ \
+        src/busy_engineers_workshop/ \
         {posargs}
     {[testenv:mypy-coverage]commands}
 
@@ -79,7 +79,7 @@ commands =
     python -m mypy \
         --py2 \
         --linecoverage-report build \
-        src/reinvent_sid345/ \
+        src/busy_engineers_workshop/ \
         {posargs}
     {[testenv:mypy-coverage]commands}
 
@@ -93,7 +93,7 @@ deps =
     flake8-print>=3.1.0
 commands =
     flake8 \
-        src/reinvent_sid345/ \
+        src/busy_engineers_workshop/ \
         setup.py \
         {posargs}
 
@@ -119,7 +119,7 @@ deps =
 commands =
     pylint \
         --rcfile=src/pylintrc \
-        src/reinvent_sid345/  \
+        src/busy_engineers_workshop/  \
         setup.py \
         {posargs}
 
@@ -140,7 +140,7 @@ deps =
     black
 commands =
     black --line-length 120 \
-        src/reinvent_sid345/ \
+        src/busy_engineers_workshop/ \
         setup.py \
         test/ \
         deploy-tools/ \
@@ -200,13 +200,13 @@ basepython = python3
 deps =
     # Pull bandit from github because they haven't published 1.4.1 to pypi yet
     git+git://github.com/PyCQA/bandit.git@master
-commands = bandit -r src/reinvent_sid345/
+commands = bandit -r src/busy_engineers_workshop/
 
 # Prone to false positives: only run independently
 [testenv:vulture]
 basepython = python3
 deps = vulture
-commands = vulture src/reinvent_sid345/
+commands = vulture src/busy_engineers_workshop/
 
 [testenv:linters]
 basepython = python3
@@ -272,7 +272,7 @@ commands =
     {[testenv:build-lambda-local]commands}
     python {toxinidir}/deploy-tools/deployer.py deploy \
         --config {toxinidir}/deploy.config \
-        --lambda-zip {toxinidir}/reinvent_sid345_artifacts/reinvent_sid345_python.zip
+        --lambda-zip {toxinidir}/busy_engineers_workshop_artifacts/busy_engineers_workshop_python.zip
 
 
 [testenv:deploy-remote-build]
@@ -284,7 +284,7 @@ commands =
     {[testenv:build-lambda-remote]commands} {posargs}
     python {toxinidir}/deploy-tools/deployer.py deploy \
         --config {toxinidir}/deploy.config \
-        --lambda-zip {toxinidir}/reinvent_sid345_artifacts/reinvent_sid345_python.zip
+        --lambda-zip {toxinidir}/busy_engineers_workshop_artifacts/busy_engineers_workshop_python.zip
 
 
 [testenv:destroy]
